multi_agent_rag_streamlit/requirements_streamlit.txt

streamlit>=1.36.0
python-dotenv>=1.0.1

cassio>=0.1.6
langchain>=0.2.14
langchain-community>=0.2.12
langchain-core>=0.2.33
langchain-text-splitters>=0.2.2
langchain-huggingface>=0.0.3
langchain-groq>=0.1.9
langgraph>=0.2.35

sentence-transformers>=3.0.1
transformers>=4.44.0
tiktoken>=0.7.0
beautifulsoup4>=4.12.3
requests>=2.32.3
